pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "brassy"
version = "0.0.3"
description = "Builds release notes from yaml files for sane software development"
authors = ["Gwyn Uttmark <biosafetylevel5@gmail.com>"]

packages = [{include="brassy", from="src"}]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies] # must download to run
python = ">=3.8"
pygit2 = "*"
rich = "*"
rich-argparse = "*"
pyyaml = "*"
platformdirs = "*"
pydantic = "*"
dateparser = "*"
eval-type-backport = "*"

# Doc group
sphinx = { version = "*", optional = true }
sphinxcontrib-runcmd = { version = "*", optional = true }
sphinxcontrib-napoleon = { version = "*", optional = true }
sphinx-argparse = { version = "*", optional = true }
pinkrst = { version = "*", optional = true }
black = { version = "*", optional = true }

# Test group
pytest = { version = "*", optional = true }

# Lint group
ruff = { version = "*", optional = true }

[tool.poetry.extras]
doc = [
    "sphinx",                    # Required for building documentation
    "sphinxcontrib-runcmd",      # Required for running shell commands in documentation
    "sphinxcontrib-napoleon",    # Required for Google-style docstrings
    "sphinx-argparse",           # Required for argparse documentation
    "pinkrst",                   # for rst formatting
    "black",                     # for rst formatting
]

test = [
    "pytest",                    # for running tests
]

lint = [
    "ruff",                      # for linting
]

[tool.poetry.scripts]
brassy = "brassy.brassy:run_from_CLI"

[tool.poetry.urls]
Homepage = "https://github.com/biosafetylvl5/brassy"
Issues = "https://github.com/biosafetylvl5/brassy/issues"